+++ ideal/contrib/django/ideal_compat/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'ideal.contrib.django.ideal_compat.apps.IdealConfig'