+++ CHANGELOG.rst
@@ -2,6 +2,17 @@
 Change history
 ==============
 
+0.2.0
+=====
+
+* February 12, 2018*
+
+* Added support for Django 1.8 and 1.11.
+* Dropped support for Django 1.7 and below.
+* Added tests for Django integration.
+* Updated documentation to reflect changes.
+
+
 0.1.1
 =====